Move span processing module into separate file
